spotcheck-button-navigation-fixes
