+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "surveysparrow-react-native-sdk",
-  "version": "1.0.8-beta.5",
+  "version": "1.0.8-beta.6",
   "description": "SurveySparrow React Native SDK enables you to collect feedback from your mobile app. Embed the Classic, Chat & NPS surveys in your React Native application seamlessly with few lines of code.",
   "source": "./src/index.tsx",
   "main": "./lib/commonjs/index.js",
@@ -85,10 +85,10 @@
   "peerDependencies": {
     "react": "*",
     "react-native": "*",
-    "react-native-webview": "^13.13.2",
     "react-native-device-info": "^14.0.4",
     "react-native-safe-area-context": "^5.3.0",
-    "react-native-svg": "15.8.0"
+    "react-native-svg": "15.8.0",
+    "react-native-webview": "^13.13.2"
   },
   "workspaces": [
     "example",
@@ -185,15 +185,15 @@
     "version": "0.38.1"
   },
   "dependencies": {
+    "@react-navigation/native": "^7.0.14",
+    "@react-navigation/stack": "^7.1.1",
     "@reduxjs/toolkit": "^2.5.1",
     "axios": "^1.7.9",
     "react-native-device-info": "^14.0.4",
     "react-native-safe-area-context": "^5.3.0",
+    "react-native-svg": "15.8.0",
     "react-native-uuid": "^2.0.3",
     "react-native-webview": "^13.13.2",
-    "react-redux": "^9.2.0",
-    "react-native-svg": "15.8.0",
-    "@react-navigation/native": "^7.0.14",
-    "@react-navigation/stack": "^7.1.1"
+    "react-redux": "^9.2.0"
   }
 }
